finalized how the outputs will be split up
