+++ generator/aggregator.go
