+++ generator/reducer.go
